Fix bug (#22) to prevent AWS configuraton corruption
This fixes #22, a bug that caused the AWS credentials to become
corrupted by inserting a "[None]" section into the configuration. Since
it doesn't make sense to write to a saved profile if one isn't set, it
will just export instead.

It's likely the case people will have 'default' in use, so to prevent
clobbering we'll default to 'sts' instead. Of course, if the user wants
to specify a profile they can.

* Before this patch, if someone ran this script with no profile, the
  ~/.aws/(config|credentials) would not have been modified.
* After this patch, the AWS config will be updated.