Add use of logging instead of print statements
I had difficulty troubleshooting an issue I was experiencing using this
utility, so I added logging, to be able to debug not only this code,
but the use of its dependent libraries.

This is not feature complete, as I'm sure there is room for improvement
but it provided the necessary functionality for me to successful
diagnose and resolve my issue.