Add dual-prompt to list of challenge choices
This fixes a small issue from 77e7544 (#206) where the code for
handling the dual-prompt page was added, but the support for it
wasn't added in the challenge-selection stage. This adds the `elif`
clause neeeded to parse the dual prompt method.

(Tested against 0.0.37 and login works successfully after this
commit is applied.)

Fixes #200